requirements.txt
Flask~=3.1.0
joblib~=1.4.2
requests~=2.32.3
scikit-learn~=1.6.1
python-dotenv~=1.0.1
polars~=1.24.0